+++ packages/react-native/ReactCommon/react/utils/jsi-utils.h
